I have completed all of the methods and attributes we came up with in the during our last meeting.
Also, I came up with a way to create the external and internal
boundaries. Essentially, I just divided the gameplay up into sections,
and we randomly choose a secton, as well as a random long/lat location,
to place the ghost in. I will explain it in more detail during lab
tomorrow.